refactor(sync): consolidate room-helpers into base-sync-room
room-helpers.ts was only consumed by base-sync-room.ts and
document-room.ts. Inlining everything into base-sync-room.ts gives
each sync room a single import and co-locates types with their
definitions. createUpdateLog and createConnectionHub are now
module-local; only createAutoSaveTracker remains exported for
DocumentRoom. Added ConnectionHub type alias for the hub field.
